+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 ## **WORK IN PROGRESS**
 
 * (theimo1221) Fix for sceneSettings Persistence
+* (theimo1221) Respect maximum and minimum Time settings even for AC-Heating
+* (theimo1221) Better force controlling of AC-Devices
 
 ## 1.3.1 (2022-11-04)